# Conflicts:
#	src/main/java/deco2800/skyfall/worlds/generation/BiomeGenerator.java
#	src/main/java/deco2800/skyfall/worlds/generation/delaunay/WorldGenNode.java
#	src/main/java/deco2800/skyfall/worlds/world/World.java

+++ src/main/java/deco2800/skyfall/worlds/world/WorldParameters.java
@@ -168,7 +168,7 @@ public class WorldParameters {
      * Sets the width of the rivers
      * @param riverWidth The width of the rivers
      */
-    public void setRiverWidth(int riverWidth) {
+    public void setRiverWidth(double riverWidth) {
         this.riverWidth = riverWidth;
     }
 
@@ -176,7 +176,7 @@ public class WorldParameters {
      * Sets the beach width
      * @param beachWidth The beach width
      */
-    public void setBeachWidth(int beachWidth) {
+    public void setBeachWidth(double beachWidth) {
         this.beachWidth = beachWidth;
     }
